feat: Coordinator 패턴 적용
