+++ ProjectManager/ProjectManager/Source/Presentation/KanbanBoard/TaskCreateScene/TaskCreateViewModel.swift
@@ -35,7 +35,6 @@ final class TaskCreateViewModel {
         let contentDidEditEvent: Observable<String>
         let datePickerDidEditEvent: Observable<Date>
         let doneButtonTapEvent: Observable<Void>
-        let cancelButtonTapEvent: Observable<Void>
     }
     
     func transform(from input: Input) -> Output {
@@ -86,12 +85,6 @@ final class TaskCreateViewModel {
                     content: self.content,
                     deadLine: self.date
                 )
-            })
-            .disposed(by: disposeBag)
-        
-        input.cancelButtonTapEvent
-            .subscribe(onNext: { [weak self] _ in
-                // coordinater finish
             })
             .disposed(by: disposeBag)
     }
